Moving evaluation to its own module
